refactor: extract QueryExecutionCoordinator from MainContentCoordinator
